+++ app/src/main/java/com/example/cse682final/SignUpActivity.kt
@@ -71,7 +71,7 @@ class SignUpActivity : AppCompatActivity() {
                     display_name_signup.text.toString(),
                     email_signup.text.toString(),
                     currentUser?.uid!!.toString(),
-                    AccountInfo(income_amount_signup.text.toString().toInt())
+                    AccountInfo(income_amount_signup.text.toString().toFloat())
                 )
             )
         }
